Remove GC request in runtime benchmarks
The GC was added to help stablize the benchmarks but is acutally
causing them to be more unstable. Before this change it is highly
likely for the benchmark to run during the copying phase of the
collector with read-guards active. After this change the GC
runs rarely and the affect can be detected using statistical
analysis of the outliers.

Test: N/A - benchmark
Change-Id: Ib402c54f1619e0e1d035444a0ce00b490b4ab0f4