Fix issue with back while backstack is empty
With predictive back, it is possible to get into a scenario where the
last record is back popped off the back stack, but there is another pop
happening before that record is actually removed from the fragment
manager and we ended up attempting to start the removal of the same
record.

We should make sure that we have executed all pending transactions
before handling back presses and we should also ensure we update the
onBackPressedCallback when we null out the transitioning op.

RelNote: "Fixed an issue when using Predictive Back with fragments where
attempting to go back twice in quick succession on the last
BackStackRecord would cause a crash."
Test: Added PredictiveBackTest
Bug: 296916226

Change-Id: Ifa1a4c4f7982c76d76094c489ffb39542281af71