Try to fix some BTF2 test flakes by hiding InputConnections from the system.
Previously, any tests that intercepted an InputConnection to make their
own calls on it would end up interleaving those calls with calls made by
the system. This returns a fake, no-op InputConnection to the system
when tests intercept it, so they are guaranteed to be the only ones
making calls. This should make tests more deterministic. It also ensures
all tests correctly access input sessions on the UI thread and clean up
after themselves by encapsulating the correct InputConnection interception
behavior in a test rule.

This change also gets rid of an alternative method of accessing the
input session in BasicTextField2ImmIntegrationTest and replaces it with
the new rule.

Bug: b/276932521
Bug: b/278560997
Test: yes
Change-Id: I35cc6213579ac91c0186e88ace87277be0a99223