Safeparcelables in AppSearch - VisibilityDocument
Make VisibilityDocument and VisibilityPermissionDocument SafeParcelable,
and remove their inheritance to GenericDocument. GenericDocument will
contain a GenericDocumentParcel in the future, and it is unnecessary to
have this internal SafeParcel for VisibilityDocument. Plus, we have plan
NOT to store visibility documents in IcingLib in the future:
b/298118943.

For now, we still need to convert those two to GenericDocuments until
we completely treat them differently for b/298118943.

Bug: 275629842
Test: appsearch:appsearch:connectedCheck appsearch:appsearch-local-storage:connectedCheck
Change-Id: Ic237f2460f79d6997e8639e7a0e320cccbe30190