Ensure cancelled gestures are completed by fragments
When a predictive back gesture is cancelled, the operation never
completes and as a result, fragments are left in a transitioning state
and can never reach the settled RESUMED and CREATED states.

Once the back gestures has dispatched the cancel signal we need to make
sure that operations and fragments are moved back to the state before
the gesture was started. Specifically this means that we need to:
  1. make sure operations are ready for container changes
  2. Actually cancel any running operations to ensure they complete
  3. Ensure the fragment view is in the proper state.

RelNote: "Fixed an issue when cancelling a predictive back gesture where
fragments failed to make it to the correct Lifecycle state."
Test: Added AnimatorTest
Bug: 297379023

Change-Id: I7cffe7994cb032c14bec01eac3c4b9dd8c8fbf40