Prevent TalkBack from focusing AndroidViewHolder
TalkBack sometimes preferred to focus the AndroidViewHolder instead of
the View it contains, particularly in combination with scroll semantics.
Focusing this empty parent node is not useful, and it caused several
glitches relating to green-focus-box refresh and TalkBack-side merging
behavior.

The usual way to prevent this kind of bug is to mark Views
`importantForAccessibility="no"`, but AndroidViewHolder's delegate
forces it to be important. The solution I found is to mark it
`visibleToUser = false`, as this seems to be the only other way to
prevent focus.

(Also set an a11y classname on AndroidViewHolder for easier debugging in
the future.)

Bug: 298048116
Bug: 291869709
Test: AndroidViewTest.androidViewAccessibilityDelegate
Change-Id: I89b8b603f09b062171651c9920d10a67bd87388a