Move @param docs of internal constructor parameters to the constructor.
Those arguments are not visible to users so they should not be
documented on the class, which is visible.

Otherwise we get warnings such as:
WARN: Unable to find what is referred to by "@param bounds" in DClass ComplicationSlot. Did you make a typo? Are you trying to refer to something not visible to users? in declaration of ComplicationSlot in file $OUT_DIR/androidx/docs-tip-of-tree/build/unzippedJvmSources/androidx/wear/watchface/ComplicationSlot.kt at line 245.

Bug: none
Change-Id: I5e910047c7ba2334c002277b7cb4b975666ce900