Update Responses to use GMS toJson() methods
These update our preU responses to utilize toJson() for the PKCs that
are returned to us from the Fido library. We still use our own flows as
a fallback flow in case there are errors in this new approach.

This modifies the libs.versions.toml, as well as runs some verification
given the warnings shown in the update of the toml file.

Bug: 297098376
Test: Built and used in sample app
Change-Id: I708e3c2183f3b57595ccfa43bfe43497904bbc7c