Fix selection handles and text toolbar to follow the layout
Handles and text toolbar will both follow the text as it scrolls.
Handles will disappear as their anchor points move out of visibility.
TextToolbar's anchor point will also stick to visible bounds.
If the text toolbar can copy some text, the text toolbar will remain
visible even if the entire selection is not visible. It will remain
at the last point the text toolbar was updated to.

Additionally, some logic that depended on start/end selection anchors
only, and not on any selectables in between, has been updated to
check all selectables that are a part of the selection. This will
help be more accurate in cases where the start or end selectables
do not exist.

Test: ./gradlew :com:found:found:cAT
Test: ./gradlew :com:found:found:tDUT
Fixes: b/298701568
Bug: b/298067102
Change-Id: Iead43e55e45a65783856635b7a18337537f05262