Use backgroundScope in SingleProcessDataStore tests
Previously, this test used testScope to initialize the Datastore instances,
which is OK as long as the SingleProcessCoordinator does not have a real
updateFlow. On the contrary, the MultiProcessDataStoreSingleProcess test uses
the background scope because its IPC has a proper notification flow.

This CL updates SingleProcessCoordiator to do the same to ensure we do close
intermediate datastores' scopes.

I've also cleaned up the doTest blocks and just used vanilla runTest since
doTest doesn't do anything special.

Bug: n/a
Test: SingleProcessDataStoreTest

Change-Id: If3eb28818d49ed5874bad6eaff6ec190c99aa204