[Tooltip] Removing savedInstanceState for tooltips
Removing the use of savedInstanceState for tooltips since the changes that need to be made to support this outweigh the benefit of supporting it.
We initially used a Saver for tooltip, so that a shown tooltip would remain showing even on device rotation. However, MutatorMutex is not a saveable type.
To make it work, we would need to make MutatorMutex saveable. For now, users will just need to reopen closed tooltips on device rotation or app closure.

Bug: 299500338
Test: N/A
Relnote: remove the use of rememberSaveable for tooltips.
Change-Id: Icc131c852cc3b3c722954aecb0a002711e13ca96