Fix unnecessarily restarting input connection in BasicSecureTextField
When IME is configured as a password input, IME does not keep an internal state of the editor and sends all key presses as platform KeyEvents. In turn, all events are interpreted as coming from a Hardware Keyboard instead of something that IME is aware of. Then, the usual `notifyIME` logic is executed by `TextFieldState` to keep the IME up-to-date. This also causes a restart in input system if the content has also changed, not just selection. In short, this causes discontinuity e.g. when long pressing backspace from the Software Keyboard.

This change intercepts IME restart requests and only proceeds if the keyboard type is not password.

Test: :compose:foundation:foundation:cAT
Change-Id: I952fe33cee86a882bf823917f3e654c64304685e