Preemptively suppresses AutoboxingStateCreation issues in Glance
Something in a previous change caused this lint check to run against glance, even though it didn't have the dependency that contained the lint check (and the APIs to fix the lint check), causing failures on build server. Not sure why / how that passed presubmit, but suppressing these issues to unblock a future upgrade in any case.

Fixes: b/298019499
Test: lintDebug
Change-Id: I7f39490c968422f4c0668326bb7c45b38ca0dd97