FakeDragTest.startFakeDragWhileSettling flakiness
The test needs to be able to perform an action in the time window during
which ViewPager2 animates a page change. As a time-based test, it is
susceptible to not getting a chance to react on time, hence a retry (of
the attempt to execute the action in the window, not of a failed assertion).

Looks like it's harder to get this right on an Emulator than originally
on a physical hardware, hence increasing the retry-action count (still
cheaper than rerunning the whole test suite). If this fails, we might
have to disable the test.

Bug: 300101098
Test: FakeDragTest
Change-Id: I91419c274fdb191ee4a47a37b91f75cd5d4710aa