[BottomSheet] Fix drastic padding for 33 > APIs > 29 when using a TextField with small content
APIs below 33, won't see the drastic padding due to an interaction between imePadding and a bottom sheet that has small content. However, the bottomsheet will not adjust for the keyboard in API < 30.

Test: Adding a test for modal bottom sheet to test the imePadding when a textfield is focused within the modal bottom sheet. It currently only tests APIs > 32 because this solution removes
the imePadding for 33 > APIs > 29. Currently imePadding does not work for new windows that don't set setDecorFitsSystemWindows to false and set resize to the window for APIs < 30.
Bug: 285746907

Change-Id: Ia013ad6081835f217946f1f983d714779b657760