Transition listeners removed from parent removes from child.
Fixes: 300685848

A transition that was cloned from another transition copied
all of its listeners instead of using the same listener
list. That made it impossible to remove a listener from
the clone parent and having removed from the running transition.

This CL now references the cloned parent's transition listeners
for the purpose of notifying, while keeping an internal list
of listeners for internal use.

Test: new test

Change-Id: If5c34afce2558a718533df7073cb667bbc5c74e5