Close all datastore instances after test ends
This CL fixes a test failure where datastore instances lurking from
previous tests would try to read their lock file and throw an exception
since the file is cleaned up by the TempFolder rule.

We were creating these instances in the service lifecycle so
they would still be closing soon after the test but we need to ensure it
happens before the TempFolder rule cleans up the files, hence this
change is necessary.

This CL changes the MultiProcessTestRule to gracefully close all
subjects before returning from the runTest method.

Also made the creation of InterProcessCompletableController thread
safe. It was causing the flakes in InterProcessCompletableTest
(I was able to repro it by running it 1k times)

Fixes: 297448143
Test: MultiProcessDataStoreIpcTest (parameterized 500 times on FTL)
Change-Id: Iff01a5ff20b2f0228f02dba6378d2db12a2f2316