Set updated SessionConfig from SessionProcessor
In CameraX extensions, vendors provide implementations of
SessionProcessor that we need to interact with in order to support
extensions. Here we introduce a dedicated class -
SessionProcessorManager to handle the interactions with the
SessionProcessor.

In this CL, we retrieve the updated configurations from SessionProcessor
through SessionProcessor.initSession(), and we then set the updated
SessionConfig through:

1. UseCaseManager.refreshAttachedUseCases()
2. SessionProcessorManager.initialize()
3. SessionProcessor.initSession()
4. UseCaseManager.tryResumeUseCaseManager()

Bug: 300518262
Test: Open extensions test app
Change-Id: I3302622d24b055f3eed45070019c8ab0b6d0386c