Fix relative path filtering in FilterableAnchorTask
Playground projects have a non-standard root and prefixing all sub
projects with the playground root makes them look like they are children
of the root and gives them a path different from the path they would
get in the main build.

This CL fixes it by using the support root directory, which makes it
consistent with the main build paths.

Bug: n/a
Change-Id: I0a7cc1d7ca4384e4e8f098a380abee379a2721ad
Test: local