The story of flaky the little flake
Flaky the little flake was a flaky test that lived happily in the
realm of Compose. He loved playing "passing the test suite" with his
other test friends. Flaky meant well and tried his best to play the
game correctly, but the little flake was a bit awkward and a bit
impatient, and often missed. But little flake would always get back
up and try again.

His friends often tried to help.

- "Little flake, you are too impatient, you must take your time, wait
for the layout and the drawing to be done, so you can assert all your
pixels".

But little flake was from the land of Graphics, where color spaces
grow strong, where pixels roam free, where eons are measured
in milliseconds and eternities in L1 cache misses. So he kept playing
with this friends, sometimes making mistakes but always getting back
up and trying again.

One day, little flake was playing another game of pass the test suite
when a large, scary monster appeared on the playground.

- "Run away little flake! Hide!", said his friends, for the scary
monster was a powerful CI Build Bot who had heard about little flake's
mistakes.

And so little flake ran away. He was fast and nimble, avoiding
detection by the dreaded CI. The monster kept looking and looking, but
could not find little flake.

Afraid and all alone, little flake made a decision that would change
his life. He decided it was time for him to change his ways and sought
out the benevolent Software Engineer to ask for help.

– "Come, little flake", said the Software Engineer. "I can help you."
- "Thank you, I didn't know what else to do", replied little flake,
  holding back his tears
- "I will give you a magic potion to solve your problems", promised the
   Software Engineer. "But first, you must follow these rules:
     1. Count down your latches slowly and calmly, no more and no less
        than you need.
     2. Relayout your views, for the hole punch must be present before
        you assert.
     3. Deflate your bounds, to keep the rounding errors away."

Little flake took the potion and left, thanking the Software Engineer.
As he walked the path of git back home, the little flake kept
repeating the 3 rules the Software Engineer had given him. He then
committed and drank the potion...

To be continued.

Fixes: 303707387
Test: AndroidExternalSurfaceTest
RelNote: N/A
Change-Id: I304a1c836f2ce9974c8f8a9d4cef9b6f8445588c