Fix draw attributes not changing in BTF2
TextFieldLayoutStateCache has been greedily caching a TextLayoutResult object by checking whether any layout affecting attributes are changed. However, this presents a problem during draw phase which relies on the fact that the latest draw attributes are always present in the `layoutInput` inside the latest `textLayoutResult`.

There are two ways to approach this issue. One is to keep the `TextFieldLayoutStateCache` the same, pipe the latest TextStyle from composition to draw phase in `TextFieldCoreModifier`, then finally invalidate the draw whenever TextStyle changes. Additionally, if TextLayoutResult is read by the devs using `onTextLayout`, we would regardlessly have to update the returned value anyway to keep the `layoutInput` stable.

The other approach is presented in this CL. We don't really have to be too greedy around TextStyle changes and how it's considered a state mutation. As long as text layout is skipped if layout attributes are not changed, we can just update the TextLayoutResult with the addition of new TextStyle. This method solves both `onTextLayout` and our internal stale draw problem.

Test: :compose:found:found:cAT

Change-Id: Idc39e18ad37a3ff964c78c0eb4c1518fdc915ff2