Make webview-version instrumentation arg mandatory
If the instrumentation setup fails and don't install a new webview
version on devices, then tests that rely on later features won't be
executed. This is particularly problematic on older Android versions.

To prevent this, this test asserts that the installed version matches
the expected version.

Previously, it could be skipped if the test infra also failed to set the
parameter for webview-version, which is why this is now being made
mandatory.

Test: webkit/run_instrumentation_tests.sh
Fixes: 303582361
Change-Id: I6fbd62ecebe66876b6698e952e951338ff4a795c