Fix flaky test in BTF2 Undo Tests
The logic for following the changing end of the selection/cursor was faulty in BasicTextFieldv1. Sometimes layout is called even though there are no changes to selection, in this case we were always following the `selection.min`. This causes issues when selection is being extended in the `max` direction but layout happens twice for different reasons, then scroll immediately jumps to `min` because selection had not changed between 2 consecutive layout calls.

The fix for this also had a very simple issue. We were _following_ the index that was _followed_ before when selection did not change. But actually no selection change should mean we do not need to follow any index. This CL removes that enforced scroll update.

Also, `getCursorRectInScroller` now coerces the queried offset in between 0 and the text layout length to make sure that we never have to query a wrong offset that would cause a crash.

Bug: b/300871959
Test: BasicTextField2UndoTest
Change-Id: I83106c44b549cfaa904adc24f9e2c29a7d8527e5