Use SurfaceTexture backed Surface for drawing overlay
Currently we are using a Bitmap backed Surface for drawing overlay. As pointed out by Eddy, it's better to provide an efficient path from the beginning. We investigated the issue and believe that the most performant way of drawing overlay is using SurfaceTexture. In this CL:

1. Instead of 2D texture, create an external texture as the overlay texture in GlRenderer.
2. Update GlProgramOverlay to use the correct sampler.
3. In SurfaceProcessorImpl/Frame, lock the Surface and send the Canvas to the app.
4. SurfaceProcessorImpl blocks the current GL thread and wait for the overlay to be updated before drawing.

For details, see: go/camera-overlay-benchmark

Bug: 305758438
Test: manual test and ./gradlew bOS
Change-Id: Ie96a3d0ee0003d1a40cba2f366b9d44b5f6f0c6b