Break apart Compose benchmark runnables to avoid ANRs
Bug: 300472956
Bug: 311256618
Bug: 311257475
Test: Manual/ABTD benchmark runs

This changes ComposeBenchmarkRule and AndroidBenchmarkRule invocations
from executing setup and benchmarks in a single runnable on the UI
thread to instead work in batches on the UI thread.

Reinstates some previously ignored benchmarks which scrolled via
pointer events.

This should drastically reduce the frequency of (locally hard to
reproduce) ANRs in microbenchmarks in CI. In the future, this is a
step toward preventing callers from running measureRepeated directly
on the UI thread, as its target duration is 10 seconds - too long to
block the UI thread.

Change-Id: Ie40d84daf1f576ba31005f0b93887cd551b8a854