Do not prune semantics nodes covered by unimportant-for-a11y nodes
Prior to this patch, if a node was covered by a semantics that only had
a testTag, it would get still pruned out of the a11y tree. I have not
seen a bug reported about this yet, but this behavior has the potential
for an innocuous test-related change to cause a serious bug. For
example, if a decorative semi-transparent element covering the whole
screen gets a testTag added to it, all other semantics would get pruned.

This patch is a followup to aosp/2718633 where I first introduced a
distinction between semantics that are important for a11y and those only
used for other purposes like testing. In general, now that this
distinction exists, we should aim to make testTags mostly no-op for a11y
and consider it a bug when they cause a different behavior.

Test: testAccessibilityNodeInfoTreePruned_testTagOnlyDoesNotPrune
Change-Id: I3e8cf7cba4c45b15cec7230bc1ab46f2354c1346