Fix FragmentContainerView fragment initiation
When creating the Fragment internally as part of inflating a FragmentContainerView, the Fragment will be populated with proper states for id, host, and fragmentManager etc. right away so that these states are accessible on the Fragment's onInflate callback.

Test: ./gradlew fragment:fragment:cC
Bug: 307427423
Relnote: "When the Fragment of a FragmentContainerView is inflated, its states such as FragmentManager, Host, and id is now accessible in the onInflate callback."
Change-Id: I1e44c5dbf5f0e0c55be39489f4c973937f6910cd