Fix flaky mute state change backwards compat test.
It seems that GnssLocationProvider is taking up 10s of the test
execution time, which is causing flakiness in the
BasicCallControlsTest#testIsMuted_BackwardsCompat. This issue has
occurred on a different test already, so I suspect this is an issue with the test timeout parameter. For now, increasing the timeout to 20s resolves the flakiness but if this issue resurfaces, the timeout parameter for this test can be removed.

Bug: 307510760
Test: BasicCallControlsTest
Change-Id: I71c784bdc00125a16485ffcb973e4d02e2fd031e