Unbundled SQLite
This CL introduces a new project that will compile SQLite and
provide as a native library for all supported kotlin targets.

This does not YET include an implementation of sqlite-driver,
just the ability to load sqlite as a native library and 1 smoke
test to ensure it is loaded properly.

Since this project requires android native targets, i've also
added android native targets to kruth. Note that, tests of these
android native targets are not run yet,we need a gtest runner
equivelant for kotlin native tests that run on android. (b/310204202)

Also, there is a consideration whether we want the JVM / ART
implementations of driver (future work) to just delegate to the
native. It can simplify the code but not sure if shipping kotlin
native runtime will be a significant size cost or not, needs to
be investigated (vs writing C JNI code for driver impl).

Bug: 310204202
Bug: 304281116
Test: NativeSqliteLoaderTest

Change-Id: I26462cf1cad4035fce04969467b1f8009f5e7ba1