Cap Exchange VOIP Impl (Participant/Call Icon)
Implements extension support for Participants on the VOIP side.

At a high level, ExtensionManager takes care of registering each
of the VOIP supported extensions before the call is started. Internally,
the VOIP side will maintain the updated state for the extensions and the
ICS can subscribe to these updates (if it also supports the extension).

This currently doesn't have version handling for resolving the
compatability of extensions between the VOIP and ICS. Instead, this just
focuses on the V1 implementation.

Bug: 301151492
Test: make androidx.core.core-telecom.main
Test: VoipExtensionRegistrationTest
Test: Ran BasicCallControlsTest with adding Participant extension via
CallsManager#addCall to ensure that all tests still pass.

Change-Id: I1da3f311716b9db5c506f94c23dbfdaa5b187b99