Replace SwipeToRevealAction class with Composable
Based on the API feedback, we are removing `SwipeToRevealAction` class and
replacing it with `SwipeToRevealPrimaryAction`,
`SwipeToRevealSecondaryAction` and `SwipeToRevealUndoAction` composables.
This makes the material `SwipeToReveal` API similar to that of
foundation with support for colors and shapes along with recommendation
on setting the action composables.

Test: Existing tests
Bug: 297879615
Relnote: """
Material SwipeToReveal APIs (for Cards and Chips) now rely
on a slot based API (as recommended by Compose) instead of
data class based instances to create those slots. This is a
breaking change, please see the demo and sample code for
examples on how to use the new API.
"""

Change-Id: Ia89431e240b0602bfe08bceb660ff9ef1137d938