Ignore 3A required parameters for SessionProcessor sessions
We currently unconditionally write 3A settings to required parameters.
This causes a problem when a CameraPipe user wants to control the 3A
settings of their request (say they want to turn AE off for a burst).
Here we add a Metadata.Key in session parameters that allows CameraGraph
users to ignore 3A required parameters, and set it to true for
SessionProcessor sessions.

Bug: 300518819
Test: CameraX extensions test app
Change-Id: I3ef82d0f128acd4a4c65455fa2d5d5d8686e449a