Force .changed calls in remember with strong skipping mode
Intrinsic remember is using the same `irChanged` helper that is used for memoizing lambdas. This helper changes behavior in strong skipping mode, generating `changedInstance` calls for unstable parameters. This is incorrect for user-defined `remember` calls, as we should always compare contents of new parameters, and not just the reference.

This change restricts `changedInstance` calls to be only generated whenever the remember call is coming from lambda memoization by assigning a custom origin to the remember expression. Generally, we try to avoid custom origins, but in this case the generated expression is removed later in the pipeline.

Test: compiler tests
Change-Id: I95612f961c7d186d392a68a209de56c74a0def08