New ApproachLayoutModifierNode API
This CL introduces a new Modifier Node for approaching
the destination layout determined in the lookahead.

This Node requires signals from the users to indicate whether
the measurement and placement approaches are complete.
Such signals afford the system the opportunity to
skip the approach upon completion to optimize layout
performance.

This CL also added a new experimental API designed
for animations, the target of which is unknown
at the time of instantiation, such as size animations,
position animations etc.

Relnote: "New ApproachLayoutModifierNode API to
support creating custom approach logic in an explicit
Modifier Node.

New experimental DeferredTargetAnimation API for animations whose
target is unknown at instantiation.
"

Test: New tests added

Change-Id: I60745501487754b36b0e1986bc2bc7ecbac267e8