Temporarily remove LookaheadScope from AnimatedContent
Lookahead scope requires an additional measure/layout pass
for when UI tree changes. UI tree isn't expected to change
frequently, except for lazy scrolling. Each new item that
is brought in the view port will be added to the tree,
and therefore will have the additional measure and layout
pass.

Since scrolling is a perf critical use case, we need to
explore ways to optimize lookahead performance to avoid
adding additional work during scrolling. It may be possible
to copy the measure results between two passes when no
animation is active. Further investigations will be needed
to verify the feasibility of such a design and build it.
Until then, lookahead will be an opt-in feature.

Test: All tests pass
RelNote: N/A
Bug: 309769741
Change-Id: I648758ca0f3b8fb38f7b5df8e1f7c9f9cca7c804