Optimize OkLab and Lab color spaces
The main optimization is in OkLab, in which we can replace the expression:

x = sign(x) * abs(x).pow(1f/3f)

With the much simpler:

x = cbrt(x)

The original implementation was necessary because pow() does not accept
negative numbers, and pow(1/3) was used as a replacement for the cube
root function which was only introduced as a stable API in Kotlin 1.8.

cbrt() has the same cost as pow() but handles negative numbers
gracefully. It also allows us to remove the abs() and sign() calls.

Not a significant optimization, but it makes the code cleaner as well.

RelNote: Optimized OkLab and Lab color space conversions
Test: ColorSpaceTest (instrumented and host), ColorTest
Change-Id: Idc2e7389f602e9e10626360cf5523646724ac162