Disable lift prediction by default
Did a small study around pressure prediction in order to determine
if it would benefit from acceleration and jank; surprisingly, it
revealed that most lift predictions turn to be inaccurate, and
that adding acceleration makes things even worse. In the few
instances where it was detected properly (around 1%), the
prediction ratchet actually prevented it from being applied.

Because of this, this CL will make lift prediction configurable,
and disabled by default.

Bug: 232941452
Test: gradlew :input:input-motionprediction:test
Change-Id: If14f8c6c62e60c163477a347b3170f7515a623bc