Use pure coroutines to animate cursor instead of Compose animation APIs.
Fixes for both BTF1 and BTF2.

This change also manually observes window focus to stop the cursor
animation when the window loses focus or the screen is turned off. This
wasn't needed before because the animation framework automatically
paused animations when the window lost focus.

Bug: b/249422803
Relnote: "Cursor animation no longer requests frames between on and off
    states."
Test: CursorAnimationStateTest
Test: TextFieldCursorTest
Change-Id: Ia22537827b1735a0c2f6f5c732bebf79aaa3b773